+++ test/framework/prometheus_rule.go
@@ -44,19 +44,23 @@ func (f *Framework) MakeBasicRule(ns, name string, groups []monitoringv1.RuleGro
 
 func (f *Framework) CreateRule(ctx context.Context, ns string, ar *monitoringv1.PrometheusRule) (*monitoringv1.PrometheusRule, error) {
 	var (
-		rule *monitoringv1.PrometheusRule
-		err  error
+		rule      *monitoringv1.PrometheusRule
+		createErr error
 	)
 
-	err = wait.PollUntilContextTimeout(ctx, time.Second, time.Minute, false, func(ctx context.Context) (bool, error) {
-		rule, err = f.MonClientV1.PrometheusRules(ns).Create(ctx, ar, metav1.CreateOptions{})
-		if err != nil {
-			return false, err
+	err := wait.PollUntilContextTimeout(ctx, time.Second, time.Minute, false, func(ctx context.Context) (bool, error) {
+		rule, createErr = f.MonClientV1.PrometheusRules(ns).Create(ctx, ar, metav1.CreateOptions{})
+		if createErr != nil {
+			return false, nil
 		}
+
 		return true, nil
 	})
+	if err != nil {
+		return nil, fmt.Errorf("%w: %w", err, createErr)
+	}
 
-	return rule, err
+	return rule, nil
 }
 
 func (f *Framework) GetRule(ctx context.Context, ns, name string) (*monitoringv1.PrometheusRule, error) {
@@ -150,4 +154,28 @@ func (f *Framework) DeleteRule(ctx context.Context, ns string, r string) error {
 	}
 
 	return nil
+}
+
+func (f *Framework) WaitForRuleCondition(ctx context.Context, pr *monitoringv1.PrometheusRule, workload metav1.Object, resource string, conditionType monitoringv1.ConditionType, conditionStatus monitoringv1.ConditionStatus, timeout time.Duration) (*monitoringv1.PrometheusRule, error) {
+	var current *monitoringv1.PrometheusRule
+
+	if err := f.WaitForConfigResourceCondition(
+		ctx,
+		func(ctx context.Context) ([]monitoringv1.WorkloadBinding, error) {
+			var err error
+			current, err = f.MonClientV1.PrometheusRules(pr.Namespace).Get(ctx, pr.Name, metav1.GetOptions{})
+			if err != nil {
+				return nil, err
+			}
+			return current.Status.Bindings, nil
+		},
+		workload,
+		resource,
+		conditionType,
+		conditionStatus,
+		timeout,
+	); err != nil {
+		return nil, fmt.Errorf("prometheusRule status %v/%v failed to reach expected condition: %w", pr.Namespace, pr.Name, err)
+	}
+	return current, nil
 }